requirements.txt
tensorflow>=2.15.0
nibabel>=3.2.0
numpy>=1.23.0,<2.0
pandas>=1.5.0
Pillow>=9.0.0
scikit-learn>=1.0.0
